Move the binding and unbinding down to blockdevmgr
This seems to make sense, as the blockdevmgr also stores its encryption
key.
